add by hyb for best block
